get saved configs working for variables that have defsults, and refactor a little the common utility code in common.py
